requirements.txt
urwid==2.0.1
PyYAML==3.12
